+++ framework/src/main/groovy/org/moqui/impl/entity/EntityFacadeImpl.groovy
@@ -725,7 +725,7 @@ class EntityFacadeImpl implements EntityFacade {
                 }
             }
             // copy master nodes (will be merged on parse)
-            // TODO: check master/detail existance before append it into entityNode
+            // TODO: check master/detail existence before append it into entityNode
             for (MNode copyNode in extendEntity.children("master")) entityNode.append(copyNode)
         }
 
@@ -1002,7 +1002,11 @@ class EntityFacadeImpl implements EntityFacade {
         ed = (EntityDefinition) entityDefinitionCache.get(entityName)
         if (ed != null) return ed
         if (entityName.isEmpty()) return null
-        return loadEntityDefinition(entityName)
+        if (entityName.startsWith("DataDocument.")) {
+            return entityDataDocument.makeEntityDefinition(entityName.substring(entityName.indexOf(".") + 1))
+        } else {
+            return loadEntityDefinition(entityName)
+        }
     }
 
     // used in tools screens
@@ -1268,6 +1272,11 @@ class EntityFacadeImpl implements EntityFacade {
         // don't check entityName empty, getEntityDefinition() does it
         EntityDefinition ed = getEntityDefinition(entityName)
         if (ed == null) throw new EntityException("No entity found with name ${entityName}")
+        if (ed.isDynamicView && entityName.startsWith("DataDocument.")) {
+            // see if it happens to be a DataDocument and if so make a special find that has its conditions too
+            // TODO: consider addition condition methods to EntityDynamicView and handling this lower level instead of here
+            return entityDataDocument.makeDataDocumentFind(entityName.substring(entityName.indexOf(".") + 1))
+        }
         return ed.makeEntityFind()
     }
     @Override